LUCENE-2065: use generics throughout unit tests
git-svn-id: https://svn.apache.org/repos/asf/lucene/java/trunk@887181 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/org/apache/lucene/store/MockRAMInputStream.java
@@ -45,7 +45,7 @@ public class MockRAMInputStream extends RAMInputStream {
     // all clones get closed:
     if (!isClone) {
       synchronized(dir) {
-        Integer v = (Integer) dir.openFiles.get(name);
+        Integer v = dir.openFiles.get(name);
         // Could be null when MockRAMDirectory.crash() was called
         if (v != null) {
           if (v.intValue() == 1) {
